requirements.txt
arabic-reshaper==2.1.3
asgiref==3.2.10
awsebcli==3.20.2
botocore==1.21.53
Brotli==1.0.9
cement==2.8.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.3
ConfigArgParse==1.5.3
coverage==6.0
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.9.0
django-crum==0.7.9
django-dynamic-fixture
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
factory_boy
faker==9.5.2
Flask==2.0.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
future==0.16.0
gevent==21.8.0
geventhttpclient==1.5.2
greenlet==1.1.2
gunicorn==20.1.0
html5lib==1.1
idna==3.2
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
locust==2.2.3
MarkupSafe==2.0.1
msgpack==1.0.2
mysql-connector==2.2.9
mysqlclient==2.0.3
pathspec==0.5.9
Pillow==8.3.2
psutil==5.8.0
psycopg2==2.9.1
pycparser==2.20
PyPDF2==1.26.0
python-bidi==0.4.2
django-pytest
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.3.0
reportlab==3.6.1
requests==2.26.0
roundrobin==0.0.2
selenium==3.141.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
termcolor==1.1.0
urllib3==1.26.7
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==2.0.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
zope.event==4.5.0
zope.interface==5.4.0
